LsiModel: return all documents. The decision on which ones to show should be made one level higher up (here: Main)
Main: improve output messages and decide which documents to show in result. Fix a bug related to the value of k (must not be 0)

Parser: in case of an Exception take the whole file (improve this, see TODO message)

Util: the terms should be determined by the indexed documents, not separately!

TdmTest: adjust test and input files for the test (as currently the first comment is ignored, a fake "license comment" has been added to the files)